feat: fastfetch-style output with icons and tree connectors
- Template redesigned with Nerd Font icons and ├─/╰─ tree structure
- Network module returns structured Table (name/ipv4/ipv6/mac)
- Config: 16 icon fields with Nerd Font defaults
- Cache security: atomic write, XDG_CACHE_HOME support
- CPU usage via two-sample /proc/stat
- Custom modules: direct Command::new (no sh -c)
- All 25 tests pass, 0 clippy warnings

+++ flexfetch-core/src/modules/disk.rs
@@ -40,8 +40,8 @@ impl Module for DiskModule {
                 let parts: Vec<&str> = line.split_whitespace().collect();
                 if parts.len() >= 6 {
                     let pct = parts[4].trim_end_matches('%').parse::<u8>().unwrap_or(0);
-                    let filled = (pct / 10).min(10) as usize;
-                    let empty = (10 - filled).min(10) as usize;
+                    let filled = ((pct / 10).min(10)) as usize;
+                    let empty = (10 - filled).min(10);
                     let bar = format!("[{}{}]", "█".repeat(filled), "░".repeat(empty));
                     let entry = format!(
                         "{}: {} / {} {} {}",
